services/agent-langgraph/src/env.py
from __future__ import annotations

from pathlib import Path

from dotenv import load_dotenv


_ENV_LOADED = False


def load_env() -> None:
    global _ENV_LOADED
    if _ENV_LOADED:
        return

    repo_root = Path(__file__).resolve().parents[3]
    env_path = repo_root / ".env"
    if env_path.exists():
        load_dotenv(env_path)

    _ENV_LOADED = True
