services/agent-langgraph/requirements.txt
langgraph>=0.2.0
pydantic>=2.6.0
psycopg[binary]>=3.1.18
python-dotenv>=1.0.1
requests>=2.32.0
minio>=7.2.0
